+++ packages/fhir-jembi/test/fixtures/ndr/medication-request.js
@@ -20,11 +20,11 @@ export default {
   status: 'completed',
   intent: 'order',
   doNotPerform: true,
-  medicationReference: { reference: 'f29d782985424800b3b309b60e39c229' },
+  medicationReference: { reference: 'Medication/f29d782985424800b3b309b60e39c229' },
   subject: { reference: 'Patient/30119' },
   encounter: { reference: 'Encounter/e84781ed-5f02-40ac-8c97-e7280fb153e3' },
   reasonReference: { reference: 'observartion/arv-regimen-changed-observation' },
-  basedOn: { reference: 'beffffab-5aec-43ad-9c31-3ecb90246b82CarePlan171' },
+  basedOn: { reference: 'CarePlan/beffffab-5aec-43ad-9c31-3ecb90246b82CarePlan171' },
   dispenseRequest: {},
   meta: {
     profile: [